+++ CHANGELOG.rst
@@ -17,9 +17,17 @@ Unreleased
 ----------
 * nothing unreleased
 
-[4.19.6]
+[4.19.8]
 --------
 * feat: remove client_id and client_secret columns from ``BlackboardEnterpriseCustomerConfiguration``
+
+[4.19.7]
+--------
+* feat: schema level improvement in integrated-channels
+
+[4.19.6]
+--------
+* feat: allowing for group members removal endpoint to support remove all
 
 [4.19.5]
 --------